+++ embedded-graphics/src/primitives/line.rs
@@ -14,8 +14,6 @@ use crate::{
 ///
 /// # Examples
 ///
-/// The [macro examples](../../macro.egline.html) make for more concise code.
-///
 /// ## Create some lines with different styles
 ///
 /// ```rust