+++ embedded-graphics/src/style/primitive_style.rs
@@ -7,13 +7,13 @@ use core::convert::TryFrom;
 /// is drawn.
 ///
 /// Because `PrimitiveStyle` has the [`non_exhaustive`] attribute, it cannot be created using a
-/// struct literal. To create a `PrimitiveStyle`, use the [`primitive_style!`] macro or
-/// [`PrimitiveStyleBuilder`].
+/// struct literal. To create a `PrimitiveStyle`, the [`with_stroke`](#method.with_stroke) and
+/// [`with_fill`](#method.with_fill) methods can be used for styles that only require a stroke or
+/// fill respectively. For more complex styles, use the [`PrimitiveStyleBuilder`].
 ///
 /// [primitive]: ../primitives/index.html
 /// [`PrimitiveStyleBuilder`]: ../style/struct.PrimitiveStyleBuilder.html
 /// [`non_exhaustive`]: https://blog.rust-lang.org/2019/12/19/Rust-1.40.0.html#[non_exhaustive]-structs,-enums,-and-variants
-/// [`primitive_style!`]: ../macro.primitive_style.html
 #[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
 #[non_exhaustive]
 pub struct PrimitiveStyle<C>
@@ -145,9 +145,6 @@ where
 /// Use this builder to create [`PrimitiveStyle`]s. If any properties on the builder are omitted,
 /// the value will remain at its default value.
 ///
-/// The [`primitive_style!`] macro can also be used to create [`PrimitiveStyle`]s, but with a
-/// shorter syntax. See the [`primitive_style!`] documentation for examples.
-///
 /// # Examples
 ///
 /// ## Build a style with configured stroke and fill
@@ -194,7 +191,6 @@ where
 /// ```
 ///
 /// [`PrimitiveStyle`]: ./struct.PrimitiveStyle.html
-/// [`primitive_style!`]: ../macro.primitive_style.html
 #[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Default)]
 pub struct PrimitiveStyleBuilder<C>
 where